src/Python/requirements.txt
azure-functions
datetime
azure-identity
azure-mgmt-resource
azure-mgmt-network
ipaddress
logging